Remove shim expression data structure
Handle `i64::MIN` directly in the parser instead. Because LALRPOP
doesn't provide a way to explicitly resolve reduce-reduce conflict,
we couldn't do something like:

```
Prefix: {
    ("-" | "!") Prefix
    Value
}

Value: {
    "-" int
    int
}
```

or:

```
Prefix: {
    "-" int
    "-" | "!" Prefix
    Value
}

Value {
    int
}
```

for the same reason. So we manually split the grammar to allow or
disallow literal integers after the unary operator. It's unclear
whether this is any cleaner than duplicating `ast::Expression`,
but it was an interesting exercise in grammar construction.