Support string literals in interpreter
To do this, we lift integers into `Value`s to track provenance, so
we know which values came from operations on integers vs. labels.
The global state keeps the heap separate from static data.

Unfortunately, we don't have a way in the interpreter to have a label
point to the second element of an array, so for now all labels have
a default offset of +8 (so that accessing -8 is valid for any label).