Restructure IR for assembly tiling
1. Separate arguments and returns from temporaries.

Because of the ABI, some arguments and returns will be stored on the
stack, and not in real temporaries. It should be easier to tile
correctly after keeping them distinct.

2. Re-introduce arguments for IR return statements.

These were removed in 7f6f46f56fce8e1e16482c24c2ccbc0e6acec9f9 so that
the CFG could have a single exit block. The problem is that then the
tiler doesn't have enough information to decide whether a `_RET` should
be tiled for a callee (relative to stack pointer) or for the caller
(relative to the frame pointer). It could maybe detect based on whether
the `_RET` is the source or destination operand, but that seems error-prone.

I haven't written the dataflow analysis framework yet, but I think
an empty exit block should still work for backward analyses, so that
its predecessors (all return blocks) will be pushed onto the list first.

---

After this commit, the IR only uses the argument and return expressions
to *receive* information from the caller and any callees, respectively.
I don't think these can be removed--there needs to be some way to
propagate information from the ABI to IR temporaries. But *sending*
information, by passing arguments to callees and returning values to
the caller, can be done during tiling.