Maintain abstract assembly return invariants
Linear abstract assembly should never have a `ret` instruction, so
that we can place a single `ret` in the epilogue (after manipulating
the stack pointer) after register allocation.

But we would like a `ret` in the CFG so that we can account for the
temporaries that are being returned, and maybe enable some optimization.
This commit guarantees these invariants by pushing and popping the `ret`
from the exit block at the CFG boundary.