Implement dataflow iteration based on strongly connected components
Tarjan's SCC algorithm gives us a (reversed) topological sort of the
strongly connected components, and each component is (as far as I
can tell) in postorder.

Because the components are topologically sorted, it's impoosible for
information to propagate backwards, so we can make a single pass over
each component. We can use the old worklist algorithm within each
component until convergence.

This also fixes a bug where initializing the worklist with a single
label wasn't sufficient to propagate information through the entire
CFG.