Fix liveness analysis across calls (maybe)
Still confused as to how to handle this. My initial instinct is that
the call both uses and defines all caller saved registers, but the
problem is that the definitions are meaningless (clobbered).

Also not sure if we need to save and restore caller-saved registers
when tiling a call? I don't think so, because the live ranges would
still be the the same (plus one extra `mov temp, register`).

Per https://engineering.purdue.edu/~milind/ece573/2014fall/project/step7/step7.html:

> CALL instructions require special care. Because we do not analyze
> liveness across functions, we must make conservative assumptions
> about what happens function calls. In particular, we GEN any
> variables that may be used, and KILL any variables that must be used.
> The GEN set for any CALL instruction therefore contains all
> global variables, while the KILL set is empty.