Hide `lir::Statement::lower` from non-CFG callers
Right now, it's still possible to construct an illegal `lir::Function<lir::Fallthrough>`,
either directly through the enums or through the interface that
the `cfg::Function` trait exposes.

It may be possible to exclude the latter by splitting off a private
trait or something, but this is fine for now.