Fix termination condition for dataflow worklist algorithm
Before, this was comparing the output set before and after transfering
across an instruction. There are two misunderstandings here: one is
that the transfer function is not monotonic (due to set subtraction
for def/kill), so using the length to compare is not correct. The other
is that we actually want to compare the output set to the last time
we computed it at this program point, not to before the transfer.

This commit fixes the algorithm to (a) compare block-level inputs and
outputs, and (b) compare them to the previous worklist visit, not the
current visit. Since we were already cloning `input` and dropping the
old `output`, it's not much more expensive to keep it around for
comparison first.