Generate tighter loops
When turning the CFG back into a linear sequence, we try to put the
fallthrough label immediately after a conditional jump. This means that
when emitting loops in IR, we should put the loop body in the false branch,
so it won't be placed at the end.

This also has the benefit of greatly decreasing the live range for most loop
temporaries, so we can spill much less often in the linear scan allocator.