Implement dead code elimination
Just the minimum working implementation. It would be nice to make it
generic over `T: cfg::Function`, so we can run it on LIR as well.
There's going to be duplicate work done since we're going to call it
from the linear scan register allocator, but maybe some of it can be
reduced later.