Avoid unnecessary cloning in API
Functions that cross type boundaries (e.g. `hir::Function` to `lir::Function`)
need to construct the new data from scratch, but in some cases, like
constructing and destructing the CFG, we can reuse the old data.
Then it should be up to the caller to clone if necessary.

Also add distinct `map` and `map_ref` functions for convenience.