Reuse existing analysis results
This is more efficient at the expense of trusting the caller to pass
an analysis that matches the CFG, and furthermore that the analysis
is not outdated (e.g. after an optimization pass).

I tried to experiment with generativity and invariant lifetimes to
tie together a CFG with its analyses, but it's pretty clunky with
interleaving CFG construction, destruction, and mutation, so it's
probably not going to work out :(

- https://lord.io/lifetimes-as-tokens/
- https://raw.githubusercontent.com/Gankro/thesis/master/thesis.pdf