Switch to two-operand form of `imul`
This allows us to avoid the fixed `rax` and `rdx` registers more
often, at the cost of some implementation complexity: the two-operand
form cannot take a memory operand as the destination, so we need
special logic to handle that.

Currently, the register allocators assumes that two spill registers
are sufficient to handle all cases, but with `imul`'s restriction,
we actually could need a third:

```text
imul [t0 + t1], t2
```

This requires one register for the memory address `t0 + t1`, one register
for `t2`, and one register for shuttling the product in and out of
memory. If we implement `xchg`, then we can do this with two spill
registers by exploiting the commutativty of `imul`, but right now we'll
just crash.

We're not actually generating any code of this form right now, since
we always tile the LHS into a fresh temporary, but that might change
if we tile directly into a destination.