Fix bugs in partial redundancy elimination
It's possible for an expression to have a latest placement at the
end of a basic block: before the terminator statement, and before
it's invalidated by the merge at the beginning of the following block.

This commit fixes the optimization pass to look for such expressions
and assign them to temporaries, where before we were only looking for
expressions to cache at their use sites. As far as I understand, these
should be the only two cases we can have something in the `latest`
set, because postponable is only killed through uses and merges.

The following example demonstrates when we would need this:

```
(LABEL enter)
  (MOVE (TEMP a) (CONST 1))
  (MOVE (TEMP b) (CONST 2))
  (MOVE (TEMP c) (CONST 0))
  (JUMP loop)
(LABEL loop)
  (CJUMP (GE (TEMP c) (ADD (TEMP a) (TEMP b))) exit false)
(LABEL false)
  (MOVE (TEMP c) (ADD (TEMP c) (CONST 1)))
  (JUMP loop)
(LABEL exit)
```

Here, we want to lift the computation of `(ADD (TEMP a) (TEMP b))`
out of the loop, and right now our analysis puts it at the end of
`enter`, right before the `(JUMP loop)` statement. But since the
jump isn't a use site, we were failing to catch this.