Subsume common subexpression elimination under lazy code motion
Extend the optimization to also cache and load subexpressions
(as opposed to only complete expressions). A dead code elimination
pass may be helpful afterward to clean up any unused subexpressions,
which could happen if the outer expression is also cached.

Unfortunately, we can't tell whether an expression or a subexpression
will be used when computing the postponable sets, so we have to
conservatively assume that an expression and all of its subexpressions
are used.