Move local variables into `Emitter` as field to avoid passing
Only `emit_declaration` needs to modify the map. Also, emit strings
as arrays for now, so (a) we don't need to take `self` mutably when
emitting expressions, and (b) because it was buggy to make all strings
static data, because mutations are visible when calling the same
function with a declared string multiple times.