Fix partial redundancy elimination interactions with memory
1) The address of a memory destination is anticipated and used,
   and prevents postponing.

2) Function calls can conservatively clobber any memory operand,
   making them unavailable after a call.

With these fixes, `mandelbrot.xi` works with all optimizations turned on!

Debugging was interesting, because inserting `print` statements in
the Xi program made it work! I compiled with various optimizations
turned off until I confirmed it was just PRE causing the breakage,
and eventually narrowed it down to the `timeout` method in `MdbCalc`.
Because of the incorrect available expression analysis, the value of
the global variable `done` was hoisted out of the loop, even though
it's updated inside the `calc` function.