Update interpreter to support globals and classes
Surprisingly little needs to be changed :)

We do have to remove the workaround for static strings in the `.data` section,
which we're no longer emitting as of commit fd1c278dcd83f867f32f34c2f1513be4a9333090.
This was because strings looked like:

```text
.quad <LENGTH>
string:
.quad <WORD_0>, <WORD_1>, ...
```

So a label reference really meant offset 8 in the backing storage. If we
add static strings back at some point, we can probably just rewrite
references to them with the constant `+ 8`, e.g. `[string + 8]`, which
should have no runtime overhead.