Pass input file before linking against runtime
Not sure why this works, as all the resources I can find explain
that the linking order should be from left to right. Perhaps it's
because there's a circular dependency between `libxi` and the object
files, since the former depends on `_Imain_paai`, and `gcc` is
automatically doing another pass or something?

https://stackoverflow.com/questions/45135/why-does-the-order-in-which-libraries-are-linked-sometimes-cause-errors-in-gcc
https://eli.thegreenplace.net/2013/07/09/library-order-in-static-linking