Reuse destination for tiling expression trees
The right-hand-side of a binary expression should reuse the same
destination that the left hand side used, since the corresponding
assembly instructions are destructive. If the binary expression
is the source of a `MOVE` instruction, then we can also write
directly into the destination of the `MOVE` rather than allocating
a fresh temporary. For example:

```text
x = 1 + 2 + 3 -> mov fresh, 1
                 add fresh, 2
                 add fresh, 3
                 mov x, fresh

x = x + 1 + 2 + 3 -> add x, 1
                     add x, 2
                     add x, 3
```

This commit also includes a Python script for generating random
expression trees. Note that we rely on the Python interpreter to
detect and warn about division by 0 errors, in which case we just
manually generate another random tree.