Optimistically allocate registers
If we can't produce a valid register allocation, then reserve one
more register as a shuttle until we can. This will cause the linear
register allocation algorithm to run multiple times in the worst case,
but most functions won't require any spills and can make use of the
two extra registers.

This way, we also don't have to worry about computing and maintaining
a static upper bound on the number of shuttle registers required.